+++ .github/workflows/test_and_deploy.yml
@@ -44,7 +44,7 @@ jobs:
           python -m pip install --upgrade pip
           python -m pip install -r requirements-test.txt
           python -m pip install -r requirements.txt
-          python -m pip install -r cerulean_cloud/cloud_run_offset_tiles/requirements.txt
+          python -m pip install -r cerulean_cloud/cloud_run_infer/requirements.txt
           python -m pip install -r cerulean_cloud/cloud_run_orchestrator/requirements.txt
           python -m pip install -r cerulean_cloud/cloud_run_tipg/requirements.txt
           python -m pip install -r cerulean_cloud/titiler_sentinel/requirements.txt
