+++ Dockerfiles/Dockerfile.cloud_run_offset
@@ -4,7 +4,7 @@ WORKDIR /app
 
 SHELL ["/bin/bash", "-o", "pipefail", "-c"]
 
-COPY cerulean_cloud/cloud_run_offset_tiles/requirements.txt ./requirements.txt
+COPY cerulean_cloud/cloud_run_infer/requirements.txt ./requirements.txt
 
 RUN pip install \
       --no-input \
@@ -45,4 +45,4 @@ ENV PYTHONUNBUFFERED=1
 # - WEB_CONCURRENCY Number of worker processes. Defaults to 1
 #   For environments with multiple CPU cores, increase the number of workers
 #   to be equal to the cores available.
-ENTRYPOINT [ "python", "-m", "uvicorn", "cerulean_cloud.cloud_run_offset_tiles.handler:app", "--host", "0.0.0.0", "--proxy-headers" ]
+ENTRYPOINT [ "python", "-m", "uvicorn", "cerulean_cloud.cloud_run_infer.handler:app", "--host", "0.0.0.0", "--proxy-headers" ]
