+++ cerulean_cloud/cloud_run_infer/README.md
@@ -10,7 +10,7 @@ docker build -f Dockerfiles/Dockerfile.cloud_run_offset -t gcr.io/cerulean-33811
 # Debug image
 
 ```shell
-PORT=8080 && docker run --rm -p 8080:${PORT} -e UVICORN_PORT=${PORT} --name cloud_run_offset_tiles gcr.io/cerulean-338116/cr-infer-image
+PORT=8080 && docker run --rm -p 8080:${PORT} -e UVICORN_PORT=${PORT} --name cloud_run_infer gcr.io/cerulean-338116/cr-infer-image
 ```
 
 # Extract the SBOM
@@ -18,5 +18,5 @@ PORT=8080 && docker run --rm -p 8080:${PORT} -e UVICORN_PORT=${PORT} --name clou
 Start the container as described in [Debug image](#debug-image) and in another shell execute
 
 ```shell
-docker cp cloud_run_offset_tiles:/app/sbom.xml .
+docker cp cloud_run_infer:/app/sbom.xml .
 ```
