+++ cerulean_cloud/cloud_run_infer/schema.py
