+++ images/cloud_run_images.py
@@ -55,7 +55,7 @@ repository = gcp.artifactregistry.Repository(
 registry_url = pulumi.Output.concat(
     "europe-west1-docker.pkg.dev/", gcp.config.project, "/", repository.repository_id
 )
-cloud_run_offset_tile_image_url = registry_url.apply(
+cloud_run_infer_image_url = registry_url.apply(
     lambda url: f"{url}/{construct_name_images('cr-infer-image')}"
 )
 cloud_run_orchestrator_image_url = registry_url.apply(
@@ -69,16 +69,16 @@ registry_info = None  # use gcloud for authentication.
 model_weights = get_file_from_gcs(
     weights_bucket,
     weights_name,
-    out_path="../cerulean_cloud/cloud_run_offset_tiles/model/model.pt",
+    out_path="../cerulean_cloud/cloud_run_infer/model/model.pt",
 )
-cloud_run_offset_tile_image = docker.Image(
+cloud_run_infer_image = docker.Image(
     construct_name_images("cr-infer-image"),
     build=docker.DockerBuildArgs(
         context="../",
         dockerfile="../Dockerfiles/Dockerfile.cloud_run_offset",
         target="final",
     ),
-    image_name=cloud_run_offset_tile_image_url,
+    image_name=cloud_run_infer_image_url,
     registry=registry_info,
 )
 cloud_run_orchestrator_image = docker.Image(
