+++ test/test_cerulean_cloud/test_cloud_run_offset_tiles.py
@@ -8,7 +8,7 @@ import torchvision  # noqa necessary for torch.jit.load of icevision mrcnn model
 from rasterio.plot import reshape_as_raster
 
 import cerulean_cloud.models as models
-from cerulean_cloud.cloud_run_offset_tiles.schema import InferenceInput, PredictPayload
+from cerulean_cloud.cloud_run_infer.schema import InferenceInput, PredictPayload
 from cerulean_cloud.tiling import TMS
 from cerulean_cloud.titiler_client import TitilerClient
